% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_lr_final.R
\name{forest_metalr}
\alias{forest_metalr}
\title{Forest plot for likeliihood ratio based meta-analysis.}
\usage{
forest_metalr(metalr_obj)
}
\arguments{
\item{metalr_obj}{An abject from the metalr functions metalr_or()
or metalr_rr(). The metalr object is a list of results computed
 by the metalr functions which includes a dataframe of mle of the
 effect estimates and their corresponding 95\% CIs and ICIs.
 See the example below.}
}
\value{
Returns a forest plot of the 95\% CIs and 95\% ICIs.
}
\description{
The function plots confidence limits of traditional 95\% CIs
and 95\% ICIs for the studies included in the meta-analysis as well as confidence bars
associated with the overall effect estimates.
}
\examples{
\dontrun{
data("statindata")  #statin potency and acute kidney injury dataset
# the metalr object
metalr_obj<-metalr_or(idata=statindata[,2:5],refval=0,num_iter=3000,increm=0.001,method = "random")
#forest plot of the metalr object
forest_metalr(metalr_obj)
}

}
