% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_lr_final.R
\name{ici.or}
\alias{ici.or}
\title{95\% Intrinsic Confidence Interval (ICI) for Odds Ratio (OR) in observational studies.}
\usage{
ici.or(idata)
}
\arguments{
\item{idata}{Vector of length 4: cases for treatment A, controls for treatment A,
cases for treatment B and control for treatment B.}
}
\value{
OR : MLE estimate of the odds ratio

llci : Lower 95\% traditional confidence limit

ulci : Upper 95\% traditional confidence limit

llici : Lower 95\% intrinsic confidence limit

ulici : Upper 95\% intrinsic confidence limit
}
\description{
Calculates traditional and intrinsic confidence intervals for
 odds ratio from an observational study.
}
\examples{
\dontrun{
data("statindata") # statin potency and acute kidney injury data
ici.or(idata = statindata[1,2:5]) # ICI for study
}

}
\references{
Dormuth, Colin R., Kristian B. Filion, and Robert W.
Platt. "Likelihood ratio meta-analysis: New motivation and approach
 for an old method." Contemporary clinical trials 47 (2016): 259-265.
}
\keyword{ICIs}
\keyword{Likelihood-ratio}
\keyword{Meta-analysis}
