% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_lr_final.R
\name{ici.rr}
\alias{ici.rr}
\title{95\% Intrinsic confidence intervals for Rate Ratios (RR) in epidemiological studies.}
\usage{
ici.rr(cases, patients, person_yrs)
}
\arguments{
\item{cases}{The number of individuals affected by the condition}

\item{patients}{The total number of individuals in the study}

\item{person_yrs}{The amount of time the patients were followed during the study}
}
\value{
RR MLE estimate of the rates ratio

llci Lower 95\% traditional confidence limit

ulci Upper 95\% traditional confidence limit

llici Lower 95\% intrinsic confidence limit

ulici Upper 95\% intrinsic confidence limit
}
\description{
Calculates 95\% traditional confidence limits and 95\% intrinsic confidence
intervals for rate ratio from epidemiological studies.
}
\examples{
\dontrun{
# Clopidogrel vs Aspirin trial dataset
cases<-c(939,1021)
person_yrs<-c(17636,17519)
patients<-c(9599,9586)
ici.rr(cases, patients, person_yrs)
}

}
\references{
Dormuth, Colin R., Kristian B. Filion, and Robert W.
Platt. "Likelihood ratio meta-analysis: New motivation and approach
 for an old method." Contemporary clinical trials 47 (2016): 259-265.
}
\keyword{Likelihood-ratio}
\keyword{Meta-analysis}
