% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_lr_final.R
\name{metalr_rr}
\alias{metalr_rr}
\title{Likelihood ratio meta-analysis for combining rate ratios in fixed
and random effects meta-analyses.}
\usage{
metalr_rr(idata, refval, num_iter, increm, method = "random")
}
\arguments{
\item{idata}{A dataframe of atleast 4 columns of: cases for treatment A, cases for
treatment B, person time for treatment A and person time for treatment B.}

\item{refval}{The reference value for the log of the alternate hypothesis}

\item{num_iter}{The number of iterations or steps from the alternate hypothesis}

\item{increm}{The quantity of increments of the refval upto the number of iterations}

\item{method}{The meta-analytic method i.e. fixed or random effect method.}
}
\value{
Total_RE A dataframe of total effect estimate from meta analysis,
 the 95\% CIs and intrinsic CIs.

Tausq Measure of heterogeneity between the studies used in the
meta-analysis in random effect meta-analysis.

Isq The I^2 statistics

meta_results  Dataframe effect estimates from all the studies,
 the 95\% confidence limits and the 95\% intrinsic confidence limits.
}
\description{
Based on the method proposed by Dormuth et al, 2016,
the function estimates traditional 95\% confidence intervals and intrinsic
confidence intervals for combined effect estimates (RR) in meta-analysis.
 It also returns an estimate of heterogeneity accross studies as well as
 Isq statistics in random meta-analysis.
}
\examples{
\dontrun{
Random dataset
data("sample_metarr_data")
metalr_rr(idata=sample_metarr_data,refval=0,num_iter=3000,increm=0.001,method = "random")
}

}
\references{
Dormuth, Colin R., Kristian B. Filion, and Robert W.
Platt. "Likelihood ratio meta-analysis: New motivation and approach
for an old method." Contemporary clinical trials 47 (2016): 259-265.
}
\keyword{Likelihood-ratio}
\keyword{Meta-analysis}
